src/requirements.txt
django
psycopg2-binary
djangorestframework
django-debug-toolbar